move monitor into its own file
