+++ src/pkg/cli/tailAndMonitor.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"io"
-	"sync"
 	"time"
 
 	"connectrpc.com/connect"
@@ -14,47 +13,6 @@ import (
 	"github.com/DefangLabs/defang/src/pkg/term"
 	defangv1 "github.com/DefangLabs/defang/src/protos/io/defang/v1"
 )
-
-func Monitor(ctx context.Context, project *compose.Project, provider client.Provider, waitTimeout time.Duration, deploymentID string, watchCallback func(*defangv1.SubscribeResponse, *ServiceStates) error) (ServiceStates, error) {
-	if deploymentID == "" {
-		panic("deploymentID must be a valid deployment ID")
-	}
-	if waitTimeout > 0 {
-		var cancelTimeout context.CancelFunc
-		ctx, cancelTimeout = context.WithTimeout(ctx, waitTimeout)
-		defer cancelTimeout()
-	}
-
-	svcStatusCtx, cancelSvcStatus := context.WithCancelCause(ctx)
-	defer cancelSvcStatus(nil)
-
-	_, computeServices := splitManagedAndUnmanagedServices(project.Services)
-
-	var (
-		serviceStates ServiceStates
-		cdErr, svcErr error
-	)
-	wg := &sync.WaitGroup{}
-	wg.Add(2)
-
-	go func() {
-		defer wg.Done()
-		serviceStates, svcErr = WatchServiceState(svcStatusCtx, provider, project.Name, deploymentID, computeServices, watchCallback)
-	}()
-
-	go func() {
-		defer wg.Done()
-		if err := WaitForCdTaskExit(ctx, provider); err != nil {
-			cdErr = err
-			cancelSvcStatus(cdErr)
-		}
-	}()
-
-	wg.Wait()
-	pkg.SleepWithContext(ctx, 2*time.Second)
-
-	return serviceStates, errors.Join(cdErr, svcErr)
-}
 
 func TailAndMonitor(ctx context.Context, project *compose.Project, provider client.Provider, waitTimeout time.Duration, tailOptions TailOptions) (ServiceStates, error) {
 	tailOptions.Follow = true
